# Documentation/swagger .yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: "Internal Job Posting"
  description: "This is a Internal Job Posting application."
  contact: 
    name: "THIS"
    url: "https://www.torryharris.com/"
    email: "thbs@gmail.com"
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
host: localhost:8080    
schemes:
- "https"
- "http"    
paths:
...
  /admin:
    get:
      summary: list job
      description: Provides job list
      operationId: getjobs
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: job details
          required: true
          schema:
            $ref: '#/definitions/viewJob'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/viewJob'
...
  /admin/edit:
    post:
      summary: edit job details and add to db
      description: 'edit job details and add to db'
      operationId: editjobaddtodb
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Enter jobname
          required: true
          schema:
            items:
              $ref: '#/definitions/editJob'
      responses:
        '200':
          description: OK
...
definitions:
...
  viewJob:
    type: object
    properties:
      jobId:
        type: integer
        format: int64
      jobname:
        type: string
      exp:
        type: integer
      level:
        type: string
      skill-set:
        type: string
      job title:
        type: string
      vacancy:
        type: integer
      job description:
        type: string
...
  editJob:
    type: object
    properties:
      jobname:
        type: string
        example: java developer
      exp:
        type: integer
        example: '3'
      level:
        type: string
        example: D3
      skill-set:
        type: string
        example: java
      vacancy:
        type: integer
        example: '6'
      job title:
        type: string
      job description:
        type: string
        example: front end and backend developer
...